+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+### Fixes
+
+- SaveableNavStack - `isRecordReachable()` off-by-one where depth of 0 didn't check the current record
+- SaveableBackStack — Update `isRecordReachable()` to match SaveableNavStack behaviour
+
 0.33.1
 ------